requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-contrib-comments>=2.0,<3.0
django-comments-xtd>=2.0,<3.0
django-celery-email>=3.0,<4.0
redis>=3.2,<4.0
django-autoslug>=1.9.8,<2.0
django-tagulous>=1.1.0,<2.0